+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this repository adheres to [Semantic Versioning](https://semver.org/spec/v2.
 
 ## Unreleased
 
+- Rename `splunk.distro.version` to `telemetry.distro.version`.
+- Declare compatibility with GDI spec 1.7.0 
+  [#2148](https://github.com/signalfx/splunk-otel-java/issues/2148)
+
 ## v2.11.0 - 2025-01-09
 
 ### General